Added exceptions to the Students and fixed them on the class.

+++ Logic/Managers/ClassManager.cs
@@ -1,6 +1,7 @@
 ﻿using AutoMapper;
 using Data;
 using Data.Models;
+using Logic.Exceptions;
 using Logic.Managers.Interfaces;
 using Logic.Models;
 
@@ -12,15 +13,33 @@ namespace Logic.Managers
         private readonly IMapper _mapper = mapper;
         public ClassDto AddStudentToClass(Guid classCode, Guid studentId)
         {
+            if (_unitOfWork.StudentRepository.GetById(studentId) == null)
+            {
+                throw new NotFoundException("No Valid Student Id found");
+            };
+
+            if (_unitOfWork.ClassRepository.GetById(classCode) == null)
+            {
+                throw new NotFoundException("No Valid class code found");
+            };
             var studentList = _unitOfWork.ClassRepository.AddStudent(classCode, studentId);
             return AssembleClass(classCode, studentList);
         }
 
         public ClassDto Create(ClassDto newClass)
         {
+            if (newClass == null)
+            {
+                throw new BadRequestException("Fields should not be empty");
+            }
+            if (String.IsNullOrEmpty(newClass.Title)|| String.IsNullOrEmpty(newClass.Description))
+            {
+                throw new BadRequestException("Title and Description Fields cannot be empty");
+            }
+
             Class creatingClass = new()
             {
-                ClassCode = newClass.ClassCode,
+                ClassCode = new Guid(),
                 Title = newClass.Title,
                 Description = newClass.Description,
                 Students = []
@@ -28,9 +47,13 @@ namespace Logic.Managers
            return _mapper.Map<ClassDto>(_unitOfWork.ClassRepository.Create(creatingClass));
         }
 
-        public ClassDto Delete(Guid deletedClass)
+        public ClassDto Delete(Guid classCode)
         {
-           return _mapper.Map<ClassDto>(_unitOfWork.ClassRepository.Delete(deletedClass));
+            if (_unitOfWork.ClassRepository.GetById(classCode) == null)
+            {
+                throw new NotFoundException("No Valid class code found");
+            };
+            return _mapper.Map<ClassDto>(_unitOfWork.ClassRepository.Delete(classCode));
         }
 
         public IEnumerable<ClassDto> GetAll()
@@ -46,23 +69,49 @@ namespace Logic.Managers
 
         public ClassDto GetById(Guid classCode)
         {
+            if (_unitOfWork.ClassRepository.GetById(classCode) == null)
+            {
+                throw new NotFoundException("No Valid class code found");
+            };
             return AssembleClass(classCode);
         }
 
         public IEnumerable<StudentDto> GetStudentsForClass(Guid classCode)
         {
+            if (_unitOfWork.ClassRepository.GetById(classCode) == null)
+            {
+                throw new NotFoundException("No Valid class code found");
+            };
             var studentList = _unitOfWork.ClassRepository.GetStudents(classCode);
             return GetStudents(studentList);
         }
 
         public ClassDto RemoveStudentFromClass(Guid classCode, Guid studentId)
         {
+            if (_unitOfWork.ClassRepository.GetById(classCode) == null)
+            {
+                throw new NotFoundException("No Valid class code found");
+            };
+
+            
             var studentList = _unitOfWork.ClassRepository.RemoveStudent(classCode, studentId);
            return AssembleClass(classCode, studentList);
         }
 
         public ClassDto Update(ClassDto newClass)
         {
+            if (newClass == null)
+            {
+                throw new BadRequestException("Fields should not be empty");
+            }
+            if (_unitOfWork.ClassRepository.GetById(newClass.ClassCode) == null)
+            {
+                throw new NotFoundException("No Valid class code found");
+            };
+            if (String.IsNullOrEmpty(newClass.Title) || String.IsNullOrEmpty(newClass.Description))
+            {
+                throw new BadRequestException("Title and Description Fields cannot be empty");
+            }
             Class classToUpdate = _unitOfWork.ClassRepository.GetById(newClass.ClassCode);
             List<Guid>classToUpdateStudents = (List<Guid>)_unitOfWork.ClassRepository.GetStudents(classToUpdate.ClassCode);
             classToUpdate.Title = newClass.Title;
